Вытаскивание полезного кода в internal
